feat(mcp): extract registry toolbox helpers into registry-toolbox-lib
Isolate diverse ranked-entry selection and toolbox planner helpers for
focused testing without changing planner or recommend behavior.
